refactor(manager): move auth service to the folder
